Move Shared.Sandbox module to Compiletime.
